Remove `husky` dependency
Now, a developer that wishes to code in only one package can open it in
their IDE without needing to run `npm install` at the root of
`jami-web`.

Caveats:
Since the dependencies for `prettier`, `eslint` and eslint plugins are declared in
the root `package.json`, you still need to run `npm install` at the root
if you wish to use it.

Change-Id: I78208f431f1216e9d8bddafd6407aabe714877d0