Fix dockerignore and improve pre-push hook
Remove some entries in dockerignore that prevented the image from
running properly.
Improve pre-push hook to prompt user to confirm that they want to push
if there are linter warnings.
Add `extract-translation` script to pre-push hook to prevent pushing if the translation files were not updated properly in the client.
Fix eslint warnings.

Change-Id: Id4491c90c182a62c1ec66a300f29f5716ea70491