Add JAMS option when creating account
With JAMS, we do not use SFL's database to store account credentials. We
rather pass the credentials directly to jami. It internally checks the
credentials with the server, and returns ERROR_GENERIC if it fails.
Multiple jami accounts can be created with the same JAMS account. That's
why I store the credentials in accounts.json too. I splitted it into two
fields: local and jams.

GitLab: #72
Change-Id: Icc925936fb47748133637837462016c4ecbbe79e