Add wrapper comp to WebRtcProvider, CallProvider
In `CallProvider` and `WebRtcProvider`, add a wrapper component that show a loading page if certain variables aren't set (webSocket, webRtcConnection).
This makes it so that we do not need to check that they are undefined in the actual code.

In `WebRtcProvider`, put some listeners in their separate `useEffect` to avoid re-binding them unecessarily.

Change-Id: I25f0ea57455bf3eb7705b69793fe3f2949c52916